+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+## [0.0.40] - 2024-03-31
+### Added
+- `RedisQueuedLocks::Client#clear_dead_requests` implementation;
+- Logger and instrumentation are passed everywhere where any changes in Redis (with locks and queus)
+  are expected;
+
 ## [0.0.39] - 2024-03-31
 ### Added
 - Logging: